requirements.txt
torch 
torchvision
scikit-image
matplotlib
pyyaml
tqdm